Step 2 : 컴포넌트 분리
